+++ src/meshinfo.cpp
@@ -2,7 +2,7 @@
 
 namespace lwpp
 {
-	int MeshInfo::scanPolys(MeshPolygonScan *scanner)
+	size_t MeshInfo::scanPolys(MeshPolygonScan *scanner)
 	{
 		scanner->setMeshInfo(this);
 		return scanPolys(scanner->LWPolyScanFunc, scanner);
